refactor: move TokensRef to the sources root
