+++ src/decorators/index.ts
@@ -3,5 +3,3 @@ export { Inject } from "./inject";
 export { Injectable } from "./injectable";
 export { InjectAll } from "./injectAll";
 export { Scoped } from "./scoped";
-export type { TokensRef } from "./tokensRef";
-export { ref } from "./tokensRef";
