+++ src/index.ts
@@ -1,7 +1,6 @@
 export type { Container, ContainerOptions } from "./container";
 export { createContainer } from "./container";
-export type { TokensRef } from "./decorators";
-export { AutoRegister, Inject, Injectable, InjectAll, ref, Scoped } from "./decorators";
+export { AutoRegister, Inject, Injectable, InjectAll, Scoped } from "./decorators";
 export { inject, injectBy } from "./inject";
 export { injectAll } from "./injectAll";
 export { Injector } from "./injector";
@@ -19,3 +18,5 @@ export { Build, Value } from "./registry";
 export { Scope } from "./scope";
 export type { Constructor, Token, TokenList, Tokens } from "./token";
 export { Type } from "./token";
+export type { TokensRef } from "./tokensRef";
+export { ref } from "./tokensRef";
