+++ src/metadata.ts
@@ -1,8 +1,8 @@
-import type { TokensRef } from "./decorators";
 import type { ClassProvider } from "./provider";
 import type { Dependencies } from "./registry";
 import type { Scope } from "./scope";
 import type { Constructor } from "./token";
+import type { TokensRef } from "./tokensRef";
 
 export interface Metadata<This extends object = any> {
   autoRegister?: boolean;
